better support for the default Nim allocator
Each thread has it's own heap, and then there's also the shared heap.
An `ActorObject` was created on the thread-local heap (`create`), but
deallocated on the shared one `deallocShared`.

Using `createShared` for allocating the actor makes it work for Nim
before version 2. At the moment, an allocator bug with Nim 2.0 makes it
impossible to allocate and deallocate an `ActorObject` on different
threads.